sequel to r8889 - eliminate type mismatch warning
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.5@8902 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ gftool/gfcksum/gfcksum.c
@@ -40,7 +40,7 @@ display_cksum(char *p, struct gfs_stat_cksum *c)
 	if (c == NULL || c->len == 0)
 		printf("%s: no checksum\n", b);
 	else
-		printf("%.*s (%s) %d %s\n", c->len, c->cksum, c->type,
+		printf("%.*s (%s) %d %s\n", (int)c->len, c->cksum, c->type,
 		    c->flags, b);
 }
